requirements/base.txt
# private packages example
# ./packages/example.tgz

starlette==0.13.3
./packages/python_gql-0.0.14-py3-none-any.whl
./packages/starlette_graphql-0.1.2-py3-none-any.whl
pydantic==1.5.1
graphql-core==3.1.0

# deploy
uvicorn==0.11.3
gunicorn==20.0.4

# database
alembic==1.4.2
SQLAlchemy==1.3.15
databases==0.3.2
asyncpg==0.20.1

# only for alembic
psycopg2-binary==2.8.4


python-socketio==4.6.0
